+++ mapsforge-samples-android/src/main/java/org/mapsforge/samples/android/MultiMapLowResWorld.java
@@ -79,7 +79,7 @@ public class MultiMapLowResWorld extends DefaultTheme {
 //            }
         };
         worldMapFile.setPriority(-1);
-        multiMapDataStore = new MultiMapDataStore(MultiMapDataStore.DataPolicy.RETURN_ALL);
+        multiMapDataStore = new MultiMapDataStore(MultiMapDataStore.DataPolicy.DEDUPLICATE);
         MapFile mapFile1 = (MapFile) getMapFile1();
         //mapFile1.restrictToZoomRange((byte) 8, Byte.MAX_VALUE);
         multiMapDataStore.addMapDataStore(mapFile1, true, true);