+++ docs/Changelog.md
@@ -2,8 +2,10 @@
 
 ## Next version
 
+- Multi-map deduplicate optimization [#1635](https://github.com/mapsforge/mapsforge/pull/1635)
 - Color filter theme resources [#1622](https://github.com/mapsforge/mapsforge/pull/1622)
   - `ThemeCallback.getBitmap`
+- Motorider map theme improvements [#1483](https://github.com/mapsforge/mapsforge/issues/1483)
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/mapsforge/issues?q=is%3Aclosed+milestone%3A0.24.0)
 
@@ -76,7 +78,7 @@
 
 ## Version 0.18.0 (2022-06-18)
 
-- Multiple maps: optimize deduplicate [#1288](https://github.com/mapsforge/mapsforge/pull/1288)
+- Multi-map deduplicate optimization [#1288](https://github.com/mapsforge/mapsforge/pull/1288)
 - Validate coordinates option [#1294](https://github.com/mapsforge/mapsforge/pull/1294)
   - `Parameters.VALIDATE_COORDINATES`
 - Minor improvements and bug fixes